pyproject.toml
[tool.poetry]
name = "organisation-label-automation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "organisation_label_automation"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
